♻️ Refactor: 패키지 이름 및 구조 변경
